% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseControl_AF.R
\name{CaseControl_AF}
\alias{CaseControl_AF}
\title{Method to reconstruct case and control AFs from the sample observed AF}
\usage{
CaseControl_AF(N_case, N_control, AF_population, OR)
}
\arguments{
\item{N_case}{the number of cases in the sample}

\item{N_control}{the number of controls in the sample}

\item{AF_population}{a vector with the observed sample AFs}

\item{OR}{a vector with the ORs 
*NOTE* Make sure the vectors are in the same order for your variants}
}
\value{
returns a dataframe with two columns (AF_case, AF_control) and rows 
equal to the number of variants
}
\description{
Method to reconstruct case and control AFs from the sample observed AF
}
