% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseControl_SE.R
\name{CaseControl_SE}
\alias{CaseControl_SE}
\title{CaseControl_SE}
\usage{
CaseControl_SE(OR, SE, N_case, N_control)
}
\arguments{
\item{OR}{a vector of odds ratios}

\item{se}{a vector of standard errors for the OR calculation *make sure the indices match between the two vectors}

\item{N_case}{an integer of the number of Case individuals}

\item{N_control}{an integer of the number of Control individuals}
}
\value{
a dataframe with 3 columns: MAF_case, MAF_control, MAF_pop for the estimated AFs for each variant which are the rows
}
\description{
This is a function to derive the case and control AFs from GWAS summary statistics when 
the user has access to the whole sample AF, the sample sizes, and the OR (or beta).
If user has SE instead of sample AF use \link[CaseControlAF]{CaseControl_AF}
This code uses the GroupFreq function adapted from C from \url{https://github.com/Paschou-Lab/ReAct/blob/main/GrpPRS_src/CountConstruct.c}
}
