% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_GWAS.R
\name{CC_GWAS}
\alias{CC_GWAS}
\title{CC_GWAS}
\usage{
CC_GWAS <- function(trait1_AF_case, trait2_AF_case, trait1_N_case, trait2_N_case, trait1_AF_control, trait2_AF_control, trait1_N_control, trait2_N_control)
}
\arguments{
\item{trait1_data}{a dataframe of the data for trait 1 containing: chromosome, position, AF in cases, AF in controls, rsID}

\item{trait2_data}a dataframe of the data for trait 2 containing: chromosome, position, AF in cases, AF in controls, rsID}

\item{trait1_N_case}{an integer of the number of Case individuals for trait 1}

\item{trait1_N_control}{an integer of the number of Control individuals for trait 1}

\item{trait2_N_case}{an integer of the number of Case individuals for trait 2}

\item{trait1_N_control}{an integer of the number of Control individuals for trait 2}

\item{AF_case_colname}{a string with the name of the column containing the AF data for cases (must be the same for trait1_data and trait2_data). Default: "AF_case"}

\item{AF_control_colname}{a string with the name of the column containing the AF data for controlss (must be the same for trait1_data and trait2_data). Default: "AF_control"}
                                                                                                     
\item{chromosome_colname}{a string with the name of the column containing the chromosome data (must be the same for trait1_data and trait2_data). Default: "chr"}

\item{position_colname}{a string with the name of the column containing the position (bp) data (must be the same for trait1_data and trait2_data). Default: "pos"}

\item{A1_colname}{a string with the name of the column containing the allele 1 data (must be the same for trait1_data and trait2_data). Default: "A1"}  

\item{A2_colname}{a string with the name of the column containing the allele 2 data (must be the same for trait1_data and trait2_data). Default: "A2"} 

\item{rsid_colname}{a string with the name of the column containing the rsID data (must be the same for trait1_data and trait2_data). Default: "RSID"} 

\item{id_colname}{a string with the name of the column containing the id data for each SNP (chr:pos) (must be the same for trait1_data and trait2_data). Default: NA. If NA, then the method will create an id column}  
}
\value{
a dataframe with 3 columns: chr, pos, RSID, A1, A2, AF_case.trait1, AF_case.trait2, AF_control.trait1, AF_control.trait2, OR, SE, p, OR_control, SE_control. Each row contains the information for a SNP that was unique and present in both datasets. OR, SE, and p contain the results of the case-case GWAS. OR_control and SE_control contain the stratification results from running the regression on controls only. 
}
\description{
This is a function to run a case-case genome wide association study (CC-GWAS) when the case and control allele frequencies (AFs) are known for two traits. This function applies the framework introduced in ReACt (\url{https://doi.org/10.1038/s41598-022-12185-6}). Details on reconstructing case and control AFs using methods within this package can be found here: \link[CaseControlAF]{CaseControl_AF} or \link[CaseControlAF]{CaseControl_SE}
This function uses code adapted from C from \url{https://github.com/Paschou-Lab/ReAct/blob/main/cc_src/}
}
